Refactor - new class RecordRetriever
