+++ src/main/java/org/folio/rest/impl/CopycatImpl.java
@@ -13,6 +13,7 @@ import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import org.folio.copycat.JsonMarc;
 import org.folio.copycat.RecordImporter;
+import org.folio.copycat.RecordRetriever;
 import org.folio.rest.annotations.Validate;
 import org.folio.rest.jaxrs.model.CopyCatImports;
 import org.folio.rest.jaxrs.model.CopyCatTargetCollection;
@@ -45,88 +46,6 @@ public class CopycatImpl implements org.folio.rest.jaxrs.resource.Copycat {
   private static final String PROFILE_TABLE = "targetprofile";
   private static Logger log = LogManager.getLogger(CopycatImpl.class);
 
-  /**
-   * Construct yaz4j Query based on external identifier and query mapping.
-   * @param profile target profile
-   * @param externalId Identifier to use within query
-   * @return Query RPN Query pattern with $identifier being replaced.
-   */
-  static Query constructQuery(CopyCatTargetProfile profile, String externalId) {
-    // assuming the externalId does not have whitespace or include {}"\\ characters
-    String pqf = profile.getExternalIdQueryMap().replace("$identifier", externalId);
-    return new PrefixQuery(pqf);
-  }
-
-  static void setAuthOptions(Connection conn, String auth) {
-    if (auth != null) {
-      String[] s = auth.trim().split("\\s+");
-      conn.option("user", s[0]);
-      if (s.length == 2) {
-        conn.option("password", s[1]);
-      } else if (s.length == 3) {
-        conn.option("group", s[1]);
-        conn.option("password", s[2]);
-      }
-    }
-  }
-
-  /**
-   * Search and retrieve record.
-   * @param profile Target Profile
-   * @param externalId record identifier such as ISBN number, OCLC number.
-   * @param type render type ("json", "xml", "raw") . See
-   *             <a href="https://software.indexdata.com/yaz/doc/zoom.records.html">ZOOM_record_get</a>
-   * @return record content
-   */
-  static Future<byte[]> getRecordAsJsonObject(CopyCatTargetProfile profile, String externalId,
-                                              String type) {
-    Connection conn = new Connection(profile.getUrl(), 0);
-    conn.option("timeout", "15");
-    conn.option("preferredRecordSyntax", "usmarc");
-    setAuthOptions(conn, profile.getAuthentication());
-
-    TargetOptions targetOptions = profile.getTargetOptions();
-    if (targetOptions != null) {
-      for (Map.Entry<String, Object> entry : targetOptions.getAdditionalProperties().entrySet()) {
-        if (entry.getValue() instanceof String) {
-          conn.option(entry.getKey(), (String) entry.getValue());
-        } else if (entry.getValue() instanceof Integer) {
-          conn.option(entry.getKey(), Integer.toString((Integer) entry.getValue()));
-        } else {
-          return Future.failedFuture("Illegal options type for key " + entry.getKey()
-              + ": " + entry.getValue().getClass());
-        }
-      }
-    }
-    Query query = constructQuery(profile, externalId);
-    try {
-      conn.connect();
-      ResultSet search = conn.search(query);
-      if (search.getHitCount() == 0) {
-        return Future.failedFuture("No record found");
-      }
-      Record record = search.getRecord(0);
-      conn.close();
-      return Future.succeededFuture(record.get(type));
-    } catch (ZoomException e) {
-      conn.close();
-      return Future.failedFuture(e);
-    }
-  }
-
-  static Future<JsonObject> getRecordAsJsonObject(CopyCatTargetProfile profile, String externalId) {
-    return getRecordAsJsonObject(profile, externalId, "json")
-        .map(buf -> new JsonObject(new String(buf)));
-  }
-
-  static Future<JsonObject> getRecordAsJsonObject(CopyCatTargetProfile profile, String externalId,
-                                                  Context vertxContext) {
-    // execute in separate thread, because getRecordAsJsonObject is a blocking function.
-    return Future.future(promise0 -> vertxContext.owner().<JsonObject>executeBlocking(promise1 ->
-            getRecordAsJsonObject(profile, externalId)
-                .onComplete(record -> promise1.handle(record)), result -> promise0.handle(result)));
-  }
-
   @Validate
   @Override
   public void postCopycatImports(CopyCatImports entity, Map<String, String> okapiHeaders,
@@ -142,7 +61,8 @@ public class CopycatImpl implements org.folio.rest.jaxrs.resource.Copycat {
             return Future.failedFuture("No such targetProfileId " + targetProfileId);
           }
           CopyCatTargetProfile targetProfile = res.mapTo(CopyCatTargetProfile.class);
-          return getRecordAsJsonObject(targetProfile, entity.getExternalIdentifier(), vertxContext)
+          return RecordRetriever.getRecordAsJsonObject(targetProfile,
+              entity.getExternalIdentifier(), vertxContext)
               .compose(marc -> {
                 if (entity.getInternalIdentifier() != null) {
                   String pattern = targetProfile.getInternalIdEmbedPath();
